Python3.10+: Faster compilation with ordered-set installation
* For 3.10 or higher we now use a maintained package that works.

* Added warning if the fallback is used, as it's very slow in comparison due to
  heavy use that data structure in optimization.
