+++ requirements-devel.txt
@@ -1,9 +1,10 @@
 # Speed up Nuitka by having this available. On Windows, installing on 2.7 is not
 # working easily, and 3.10 has issues on everything but Linux, but definitely on
 # macOS and Windows.
-orderedset == 2.0.3 ; os.name != "nt" and sys.platform != "darwin"
+orderedset == 2.0.3 ; os.name != "nt" and sys.platform != "darwin" and python_version <= "3.9"
 orderedset == 2.0.3 ; os.name == "nt" and python_version >= '3.4' and python_version <= "3.9"
 orderedset == 2.0.3 ; sys.platform == "darwin" and python_version <= "3.9"
+ordered-set == 4.1.0; python_version >= "3.10"
 
 # Make testing more reliable on Python2 with timeout features added.
 subprocess32 == 3.5.4 ; python_version == '2.7'
